Refactored the health checks
